Reverted r24463 as this is needed to conserve memory, especially for a large number of Sobol' points.
The command used was:
svn merge -r24463:r24462 .

.....
  r24463 | bugman | 2014-07-05 19:06:09 +0200 (Sat, 05 Jul 2014) | 9 lines
  Changed paths:
     M /branches/frame_order_cleanup/target_functions/frame_order.py
  
  Converted all frame order target function data structures to float64.
  
  This affects the vector data structures used for the PCS numeric integration.  The float32
  structures added for speed do not actually increase the speed.  For the double rotor model, 10
  target function calls took the pcs_pivot_motion_double_rotor() function 44.936 seconds using float32
  and 44.105 seconds using all float64.  The cumulative time is also longer.  This must be because
  other parts are float64, and therefore type conversions are happening during target function calls.
.....

+++ target_functions/frame_order.py
@@ -25,7 +25,7 @@
 # Python module imports.
 from copy import deepcopy
 from math import acos, cos, pi, sin, sqrt
-from numpy import add, array, dot, float64, ones, outer, subtract, swapaxes, transpose, uint8, zeros
+from numpy import add, array, dot, float32, float64, ones, outer, subtract, transpose, uint8, zeros
 
 # relax module imports.
 from extern.sobol.sobol_lib import i4_sobol_generate
@@ -281,8 +281,8 @@ class Frame_order:
             self.paramag_unit_vect = zeros(atomic_pos.shape, float64)
             self.paramag_dist = zeros(self.num_spins, float64)
             self.pcs_const = zeros((self.num_align, self.num_spins), float64)
-            self.r_pivot_atom = zeros((self.num_spins, 3), float64)
-            self.r_pivot_atom_rev = zeros((self.num_spins, 3), float64)
+            self.r_pivot_atom = zeros((self.num_spins, 3), float32)
+            self.r_pivot_atom_rev = zeros((self.num_spins, 3), float32)
             self.r_ln_pivot = self.pivot - self.paramag_centre
 
             # Set up the paramagnetic constant (without the interatomic distance and in Angstrom units).
@@ -2030,7 +2030,7 @@ class Frame_order:
         # Initialise.
         sobol_data.model = self.model
         sobol_data.total_num = total_num
-        sobol_data.sobol_angles = zeros((m, total_num), float64)
+        sobol_data.sobol_angles = zeros((m, total_num), float32)
         sobol_data.Ri_prime = zeros((total_num, 3, 3), float64)
         sobol_data.Ri2_prime = zeros((total_num, 3, 3), float64)
